Progress from first hack night
There is correct insertion of words into the trie. Needs support for
empty transitions after the root node. E.g. if you insert 'was' and then
'waste', you will not have an empty transition available from the 's' to
indicate that 'was' is also a word. Deletion and searching still need to
be done.  I bet there's a better word list somewhere, too...